+++ lib/cubit/myBooking/my_booking_cubit.dart
@@ -58,12 +58,10 @@ class MyBookingCubit extends Cubit<MyBookingState> {
   }
 
   // 5. Rate Booking
-  Future<void> rateBooking(String bookingId, double rating) async {
+  Future<void> rateBooking(String bookingId, int rating) async {
     final currentStatus = _getCurrentStatus();
     try {
       await _repo.rateBooking(bookingId, rating);
-      // We refresh the list. If the backend moves rated items or changes their status,
-      // the new list will reflect that.
       getBookings(status: currentStatus);
     } catch (e) {
       emit(MyBookingError("Rating failed: ${e.toString()}", currentStatus));
